requirements.txt
matplotlib
py-cpuinfo
pytest
qiskit-terra
setproctitle
rich